#109 Pull tags from books coauthored by author
