+++ extension/ts/main/adapters/book/webscrapers/bookFromEdition.ts
@@ -3,7 +3,7 @@ import {getDocument} from "../../../services/finder/util/getDocument";
 import {scrapeCopy} from "./scrapeCopy";
 import {asyncCached} from "../bookCache";
 
-const getBook = async (link: string): Promise<BookRecord> => {
+const getBookFromEditionPage = async (link: string): Promise<BookRecord> => {
 	const id = new URL(link).pathname.split("/").pop();
 
 	return asyncCached(id, async () => {
@@ -17,4 +17,4 @@ const getBook = async (link: string): Promise<BookRecord> => {
 	});
 };
 
-export {getBook};
+export {getBookFromEditionPage};
