+++ extension/ts/main/adapters/book/webscrapers/bookFromWork.ts
@@ -1,6 +1,6 @@
 import {BookRecord} from "../types";
 import {getDocument} from "../../../services/finder/util/getDocument";
-import {getBook} from "./bookFromEdition";
+import {getBookFromEditionPage} from "./bookFromEdition";
 import {scrapeCopy} from "./scrapeCopy";
 import {syncCached} from "../bookCache";
 
@@ -14,11 +14,11 @@ const getCopy = (document: Document): BookRecord => {
 
 const getOtherEditions = async (yourCopyInfo: HTMLDivElement): Promise<BookRecord[]> => {
 	const otherEditionLinks = Array.from(yourCopyInfo.querySelectorAll<HTMLLinkElement>("div.copylist > p > b > a"));
-	const futureOtherEditions = otherEditionLinks.map((link) => getBook(link.href));
+	const futureOtherEditions = otherEditionLinks.map((link) => getBookFromEditionPage(link.href));
 	return Promise.all(futureOtherEditions);
 };
 
-const getBooksFromWork = async (link: string): Promise<BookRecord[]> => {
+const getBooksFromWorkPage = async (link: string): Promise<BookRecord[]> => {
 	const workPage = await getDocument(link);
 	const yourCopyInfo = workPage.querySelector<HTMLDivElement>(".qelcontent.bookinfo");
 	if (yourCopyInfo) {
@@ -30,4 +30,4 @@ const getBooksFromWork = async (link: string): Promise<BookRecord[]> => {
 	}
 };
 
-export {getBooksFromWork};
+export {getBooksFromWorkPage};
